Revert BaseTools: PYTHON3 migration
This reverts commit bb55ec4c882b542df1e7be4b919e54aa413994e1..
f7e19ed9cffb14945508681ab8ce3fe9f5bfa6bb.

Python3 migration is the fundamental change. It requires every developer
to install Python3. Before this migration, the well communication and wide
verification must be done. But now, most people is not aware of this change,
and not try it. So, Python3 migration is reverted and be moved to edk2-staging
Python3 branch for the edk2 user evaluation.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Liming Gao <liming.gao@intel.com>